treesitter.c: 'R' Bindings to the 'C' Grammar for Tree-Sitter

Provides bindings to a 'C' grammar for Tree-sitter, to be used alongside the 'treesitter' package. Tree-sitter builds concrete syntax trees for source files and can efficiently update them as files are edited.

See also

Author

Maintainer: Sounkou Mahamane Toure sounkoutoure@gmail.com

Other contributors:

Tree-sitter authors (Tree-sitter C grammar) [copyright holder]